Добавляет в provider проверку на доступность интернета

+++ src/api/provider.js
@@ -4,26 +4,54 @@ export default class Provider {
   }
 
   getDestinations() {
-    return this._api.getDestinations();
+    if (this._isOnline()) {
+      return this._api.getDestinations();
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   getOffers() {
-    return this._api.getOffers();
+    if (this._isOnline()) {
+      return this._api.getOffers();
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   getPoints() {
-    return this._api.getPoints();
+    if (this._isOnline()) {
+      return this._api.getPoints();
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   createPoint(point) {
-    return this._api.createPoint(point);
+    if (this._isOnline()) {
+      return this._api.createPoint(point);
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   updatePoint(id, data) {
-    return this._api.updatePoint(id, data);
+    if (this._isOnline()) {
+      return this._api.updatePoint(id, data);
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   deletePoint(id) {
-    return this._api.deletePoint(id);
+    if (this._isOnline()) {
+      return this._api.deletePoint(id);
+    }
+
+    return Promise.reject(`offline logic is not implemented`);
+  }
+
+  _isOnline() {
+    return window.navigator.onLine;
   }
 }
